# server/openapi.yaml
openapi: 3.0.3
info:
  title: BaudAgain BBS API
  description: |
    REST API for BaudAgain BBS - A modern AI-enhanced bulletin board system.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - Authentication endpoints: 10 requests per minute
    - Read operations: 100 requests per 15 minutes
    - Write operations: 30 requests per minute
    - Message posting: 30 messages per hour
    
  version: 1.0.0
  contact:
    name: BaudAgain BBS
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageBase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        accessLevelRead:
          type: integer
          minimum: 0
          maximum: 255
        accessLevelWrite:
          type: integer
          minimum: 0
          maximum: 255
        postCount:
          type: integer
        lastPostAt:
          type: string
          format: date-time
        sortOrder:
          type: integer
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        messageBaseId:
          type: string
          format: uuid
        subject:
          type: string
        body:
          type: string
        authorId:
          type: string
          format: uuid
        authorHandle:
          type: string
        createdAt:
          type: string
          format: date-time
        parentId:
          type: string
          format: uuid
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
paths:
...
  /message-bases:
    get:
      tags:
        - Message Bases
      summary: List message bases
      description: Get all message bases
      operationId: listMessageBases
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of message bases
          content:
            application/json:
              schema:
                type: object
                properties:
                  bases:
                    type: array
                    items:
                      $ref: '#/components/schemas/MessageBase'

    post:
      tags:
        - Message Bases
      summary: Create message base
      description: Create new message base (admin only)
      operationId: createMessageBase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                accessLevelRead:
                  type: integer
                  default: 0
                accessLevelWrite:
                  type: integer
                  default: 10
                sortOrder:
                  type: integer
                  default: 0
      responses:
        '201':
          description: Message base created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBase'
        '403':
          description: Not authorized
  /message-bases/{id}:
    get:
      tags:
        - Message Bases
      summary: Get message base
      description: Get message base details
      operationId: getMessageBase
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Message base details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBase'
        '404':
          description: Message base not found

    patch:
      tags:
        - Message Bases
      summary: Update message base
      description: Update message base (admin only)
      operationId: updateMessageBase
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                accessLevelRead:
                  type: integer
                accessLevelWrite:
                  type: integer
                sortOrder:
                  type: integer
      responses:
        '200':
          description: Message base updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBase'
        '403':
          description: Not authorized
        '404':
          description: Message base not found

    delete:
      tags:
        - Message Bases
      summary: Delete message base
      description: Delete message base (admin only)
      operationId: deleteMessageBase
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Message base deleted
        '403':
          description: Not authorized
        '404':
          description: Message base not found
  /message-bases/{baseId}/messages:
    get:
      tags:
        - Messages
      summary: List messages
      description: |
        Get messages in a message base
        
        **curl Example:**
        ```bash
        curl -X GET "http://localhost:8080/api/v1/message-bases/$MESSAGE_BASE_ID/messages?page=1&limit=50" \
          -H "Authorization: Bearer $TOKEN"
        ```
      operationId: listMessages
      security:
        - bearerAuth: []
      parameters:
        - name: baseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Messages
      summary: Post message
      description: |
        Post new message to message base
        
        **curl Example:**
        ```bash
        curl -X POST "http://localhost:8080/api/v1/message-bases/$MESSAGE_BASE_ID/messages" \
          -H "Authorization: Bearer $TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "subject": "Test Message",
            "body": "This is a test message posted via the API."
          }'
        ```
      operationId: postMessage
      security:
        - bearerAuth: []
      parameters:
        - name: baseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subject
                - body
              properties:
                subject:
                  type: string
                  maxLength: 100
                body:
                  type: string
                  maxLength: 5000
                parentId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Message posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: No write access
        '429':
          description: Rate limit exceeded
  /messages/{id}:
    get:
      tags:
        - Messages
      summary: Get message
      description: Get message details
      operationId: getMessage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Message details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Message not found
  /messages/{id}/replies:
    post:
      tags:
        - Messages
      summary: Post reply
      description: Post reply to a message
      operationId: postReply
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Parent message ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subject
                - body
              properties:
                subject:
                  type: string
                  maxLength: 200
                body:
                  type: string
                  maxLength: 10000
      responses:
        '201':
          description: Reply posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: No write access
        '404':
          description: Parent message not found
        '429':
          description: Rate limit exceeded
...